Display postal address
It's clear that there will be a lot of duplication in our presenters, so may as well just create a BaseLetterPresenter.
